requirements.txt
cryptography>=38.0.0
argon2-cffi>=21.3.0
requests>=2.28.0
fido2>=1.0.0
Flask>=2.2.0
Flask-Cors>=3.0.10
PyJWT>=2.6.0
pycryptodome>=3.16.0
pyotp>=2.8.0 